+++ .gitignore
@@ -1,0 +1,3 @@
+.vscode
+mega_sena_API
+__pycache__
