+++ app/helpers/findRecurrence.py
